requirements.txt
streamlit>=1.30
Pillow>=10.0
pypdfium2>=4.25
